Factored out CORE.currentWindow
All instances of CORE.currentWindow have been replaced with a windowID variable that is passed into the function, allowing for more custom control over window and context management.

This may produce unexpected results in places where raylib is expecting the context to match the window that is being operated on and it doesn't. I'll have to watch out for those cases, I'm sure a few are lingering around that I didn't catch. The fix for this is running SetContext(windowID) in the problem area so that everything matches.

I was going to take this refactor to the next level and also require all functions in rlgl to require a contextID but I realized that would be kind of ludicrious as it changes literally every single function and also maybe kind of bad, potentially leading us to do a lot of uncessessary context switching. We probably don't want to be switching contexts a lot, right? (I have no idea if that's bad, but it feels like it would be bad)