added & ported newest raygui.h
Ported newest version of raygui.h. Only change that needed to be made to get things compiling was fix a few warnings, renamings, and accomodating for multiple windows by putting windowID in most functions (required for getting mouse position). Unsure yet if I'm going to use this and modify it or start from scratch, but it's a good starting place.