Remove more Windows.h conflicts
for raylib standalone it makes sense to trade the simplicity with no window.h compatibility, but I'm fine with the tradeoff of puting Ray before a few functions. I considered putting Ray before every function just to have it be consistent but that would make potentially upgrading raylib in the future with newer versions an absolute bear so not gonna go that far.